livestreaming/requirements.txt
# Camera Livestreaming Dependencies

# Async HTTP server and client
aiohttp==3.9.1
aiohttp-cors==0.7.0

# WebSocket support
websockets==12.0

# MQTT Client (EMQX)
paho-mqtt==1.6.1

# Database (SQLite included in Python)
# No additional dependency needed

# Logging (included in Python)
# No additional dependency needed

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
